requirements.txt
# local package
-e .

# external requirements
pytorch-lightning==1.9.4
hydra-core